Add files via upload
This is the streamlined version of data collection where each game sequence is stored as a list within game.py and all game sequences are stored is stored as a list also within main.py and exported to a JSON file. These adjustments make sequence_trie.py obsolete. I also removed the sounds from playing. This could be further streamlined if needed by eliminating the loading of sound files on game initialization.